Fix up some unit tests to make them extensible.
