+++ .gitignore
@@ -16,3 +16,5 @@ hs_err_pid*
 
 # generated files
 target
+
+gabby.iml
